[FE] refactor: Box, BoxHeader 레이아웃 컴포넌트 구현
