+++ fe/src/components/common/box/BoxHeader.tsx
@@ -1,0 +1,22 @@
+import { useTheme } from '@emotion/react';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+export const BoxHeader: React.FC<Props> = ({ children }) => {
+  const theme = useTheme() as any;
+
+  return (
+    <div
+      css={{
+        minHeight: '64px',
+        borderBottom: `${theme.border.default} ${theme.neutral.border.default}`,
+        display: 'flex',
+        alignItems: 'center',
+      }}
+    >
+      {children}
+    </div>
+  );
+};
